rename: tab 경로 수정
